refactor: serverName Atom 분리
